+++ index.html
@@ -4,15 +4,15 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Learning with Mibubu</title>
+    <link rel="stylesheet" href="style.css" />
   </head>
   <body>
-    (open the console)
+    graph
+    <div id="mynetwork"></div>
 
-    <!-- <script src="graphs.js" type="module"></script> -->
-    <script src="graphs2.js" type="module"></script>
-    <script src="utils.js" type="module"></script>
+    <script src="https://cdn.jsdelivr.net/npm/vis-network/standalone/umd/vis-network.min.js"></script>
+    <script src="graphVis.js"></script>
     <script src="miahiGraphProbs.js" type="module"></script>
-
-    <script src="fcc10.js" type="module"></script>
+    <script src="random_probs.js" type="module"></script>
   </body>
 </html>
