+++ old_stuff/PR8/commonAncestor.js
