+++ old_stuff/PR8/multiplyNums.js
