+++ old_stuff/lesson12/timed1.js
