+++ old_stuff/lesson12/timed2.js
