+++ old_stuff/lesson12/timed3.js
