+++ old_stuff/lesson12/timed5.js
