+++ old_stuff/lesson9/timedProb3.js
