+++ style.css
@@ -1,0 +1,5 @@
+#mynetwork {
+  width: 600px;
+  height: 400px;
+  border: 1px solid lightgray;
+}
